Sync EfficientnetB4(PyTorch) code to latest
    link #IA95XR

+++ cv/classification/efficientnet_b4/pytorch/common_utils/smooth_value.py
@@ -1,6 +1,6 @@
-# Copyright (c) 2022 Iluvatar CoreX. All rights reserved.
 # Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved.
-
+# Copyright (c) 2024, Shanghai Iluvatar CoreX Semiconductor Co., Ltd.
+# All Rights Reserved.
 
 from collections import defaultdict, deque
 import datetime
@@ -37,7 +37,7 @@ class SmoothedValue(object):
         """
         if not is_dist_avail_and_initialized():
             return
-        t = torch.tensor([self.count, self.total], dtype=torch.float64, device='cuda')
+        t = torch.tensor([self.count, self.total], dtype=torch.float32, device='cuda')
         dist.barrier()
         dist.all_reduce(t)
         t = t.tolist()
